@@ -9,7 +9,7 @@
         "astrotomic/php-open-graph": "^1.0",
         "codeat3/blade-codicons": "^1.30",
         "guzzlehttp/guzzle": "^7.2",
-        "laravel/framework": "^11.0",
+        "laravel/framework": "^12.0",
         "laravel/tinker": "^2.7",
         "league/commonmark": "^2.2",
         "livewire/livewire": "^3.0",
@@ -24,7 +24,7 @@
         "laravel/sail": "^1.0.1",
         "mockery/mockery": "^1.4.4",
         "nunomaduro/collision": "^8.1",
-        "phpunit/phpunit": "^10",
+        "phpunit/phpunit": "^11",
         "spatie/laravel-ignition": "^2.0"
     },
     "autoload": {
